Add SQLw replacing placeholders in SQL statements
We add the function `SQLw()` which takes a statement and key/value
pairs as variadic arguments. Placeholders are of the format
`$(key)`. For example, a typical use-case is to substitute the
table name dynamically in queries:

    SQLw("SELECT c1 FROM $(tblName)", "tblName", "t1")

Placeholders within single or double quotes, or backticks, are ignored.

The companion function `MustSQLw` is also available: it panics instead
of reading an error.